Reorganize temporal config to middlewaresconfig
